Move test of proxyfs to test-other
