+++ tests/test_other.py
@@ -2669,6 +2669,235 @@ int main()
       out = run_js('a.out.js', engine=engine, stderr=PIPE, full_output=True)
       self.assertContained('File size: 724', out)
 
+  def test_proxyfs(self):
+    open('proxyfs_test_main.js', 'w').write(r'''
+var m0 = require('./proxyfs_test.js');
+var m1 = require('./proxyfs_test1.js');
+var m2 = require('./proxyfs_test2.js');
+
+var section;
+function print(str){
+  process.stdout.write(section+":"+str+":");
+}
+
+m0.FS.mkdir('/working');
+m0.FS.mount(m0.PROXYFS,{root:'/',fs:m1.FS},'/working');
+m0.FS.mkdir('/working2');
+m0.FS.mount(m0.PROXYFS,{root:'/',fs:m2.FS},'/working2');
+
+section = "child m1 reads and writes local file.";
+print("m1 read embed");
+m1.ccall('myreade','number',[],[]);
+print("m1 write");console.log("");
+m1.ccall('mywrite0','number',['number'],[1]);
+print("m1 read");
+m1.ccall('myread0','number',[],[]);
+
+
+section = "child m2 reads and writes local file.";
+print("m2 read embed");
+m2.ccall('myreade','number',[],[]);
+print("m2 write");console.log("");
+m2.ccall('mywrite0','number',['number'],[2]);
+print("m2 read");
+m2.ccall('myread0','number',[],[]);
+
+section = "child m1 reads local file.";
+print("m1 read");
+m1.ccall('myread0','number',[],[]);
+
+section = "parent m0 reads and writes local and children's file.";
+print("m0 read embed");
+m0.ccall('myreade','number',[],[]);
+print("m0 read m1");
+m0.ccall('myread1','number',[],[]);
+print("m0 read m2");
+m0.ccall('myread2','number',[],[]);
+
+section = "m0,m1 and m2 verify local files.";
+print("m0 write");console.log("");
+m0.ccall('mywrite0','number',['number'],[0]);
+print("m0 read");
+m0.ccall('myread0','number',[],[]);
+print("m1 read");
+m1.ccall('myread0','number',[],[]);
+print("m2 read");
+m2.ccall('myread0','number',[],[]);
+
+print("m0 read embed");
+m0.ccall('myreade','number',[],[]);
+print("m1 read embed");
+m1.ccall('myreade','number',[],[]);
+print("m2 read embed");
+m2.ccall('myreade','number',[],[]);
+
+section = "parent m0 writes and reads children's files.";
+print("m0 write m1");console.log("");
+m0.ccall('mywrite1','number',[],[]);
+print("m0 read m1");
+m0.ccall('myread1','number',[],[]);
+print("m0 write m2");console.log("");
+m0.ccall('mywrite2','number',[],[]);
+print("m0 read m2");
+m0.ccall('myread2','number',[],[]);
+print("m1 read");
+m1.ccall('myread0','number',[],[]);
+print("m2 read");
+m2.ccall('myread0','number',[],[]);
+print("m0 read m0");
+m0.ccall('myread0','number',[],[]);
+''')
+
+    open('proxyfs_pre.js', 'w').write(r'''
+if (typeof Module === 'undefined') Module = {};
+Module["noInitialRun"]=true;
+Module["noExitRuntime"]=true;
+''')
+
+    open('proxyfs_embed.txt', 'w').write(r'''test
+''')
+
+    open('proxyfs_test.c', 'w').write(r'''
+#include <stdio.h>
+
+int
+mywrite1(){
+  FILE* out = fopen("/working/hoge.txt","w");
+  fprintf(out,"test1\n");
+  fclose(out);
+  return 0;
+}
+
+int
+myread1(){
+  FILE* in = fopen("/working/hoge.txt","r");
+  char buf[1024];
+  int len;
+  if(in==NULL)
+    printf("open failed\n");
+
+  while(! feof(in)){
+    if(fgets(buf,sizeof(buf),in)==buf){
+      printf("%s",buf);
+    }
+  }
+  fclose(in);
+  return 0;
+}
+int
+mywrite2(){
+  FILE* out = fopen("/working2/hoge.txt","w");
+  fprintf(out,"test2\n");
+  fclose(out);
+  return 0;
+}
+
+int
+myread2(){
+  {
+    FILE* in = fopen("/working2/hoge.txt","r");
+    char buf[1024];
+    int len;
+    if(in==NULL)
+      printf("open failed\n");
+
+    while(! feof(in)){
+      if(fgets(buf,sizeof(buf),in)==buf){
+        printf("%s",buf);
+      }
+    }
+    fclose(in);
+  }
+  return 0;
+}
+
+int
+mywrite0(int i){
+  FILE* out = fopen("hoge.txt","w");
+  fprintf(out,"test0_%d\n",i);
+  fclose(out);
+  return 0;
+}
+
+int
+myread0(){
+  {
+    FILE* in = fopen("hoge.txt","r");
+    char buf[1024];
+    int len;
+    if(in==NULL)
+      printf("open failed\n");
+
+    while(! feof(in)){
+      if(fgets(buf,sizeof(buf),in)==buf){
+        printf("%s",buf);
+      }
+    }
+    fclose(in);
+  }
+  return 0;
+}
+
+int
+myreade(){
+  {
+    FILE* in = fopen("proxyfs_embed.txt","r");
+    char buf[1024];
+    int len;
+    if(in==NULL)
+      printf("open failed\n");
+
+    while(! feof(in)){
+      if(fgets(buf,sizeof(buf),in)==buf){
+        printf("%s",buf);
+      }
+    }
+    fclose(in);
+  }
+  return 0;
+}
+''')
+
+    Popen([PYTHON, EMCC,
+           '-o', 'proxyfs_test.js', 'proxyfs_test.c',
+           '--embed-file', 'proxyfs_embed.txt', '--pre-js', 'proxyfs_pre.js',
+           '-s', 'MAIN_MODULE=1']).communicate()
+    # Following shutil.copyfile just prevent 'require' of node.js from caching js-object.
+    # See https://nodejs.org/api/modules.html
+    shutil.copyfile('proxyfs_test.js', 'proxyfs_test1.js')
+    shutil.copyfile('proxyfs_test.js', 'proxyfs_test2.js')
+    out = run_js('proxyfs_test_main.js')
+    section="child m1 reads and writes local file."
+    self.assertContained(section+":m1 read embed:test", out)
+    self.assertContained(section+":m1 write:", out)
+    self.assertContained(section+":m1 read:test0_1", out)
+    section="child m2 reads and writes local file."
+    self.assertContained(section+":m2 read embed:test", out)
+    self.assertContained(section+":m2 write:", out)
+    self.assertContained(section+":m2 read:test0_2", out)
+    section="child m1 reads local file."
+    self.assertContained(section+":m1 read:test0_1", out)
+    section="parent m0 reads and writes local and children's file."
+    self.assertContained(section+":m0 read embed:test", out)
+    self.assertContained(section+":m0 read m1:test0_1", out)
+    self.assertContained(section+":m0 read m2:test0_2", out)
+    section="m0,m1 and m2 verify local files."
+    self.assertContained(section+":m0 write:", out)
+    self.assertContained(section+":m0 read:test0_0", out)
+    self.assertContained(section+":m1 read:test0_1", out)
+    self.assertContained(section+":m2 read:test0_2", out)
+    self.assertContained(section+":m0 read embed:test", out)
+    self.assertContained(section+":m1 read embed:test", out)
+    self.assertContained(section+":m2 read embed:test", out)
+    section="parent m0 writes and reads children's files."
+    self.assertContained(section+":m0 write m1:", out)
+    self.assertContained(section+":m0 read m1:test1", out)
+    self.assertContained(section+":m0 write m2:", out)
+    self.assertContained(section+":m0 read m2:test2", out)
+    self.assertContained(section+":m1 read:test1", out)
+    self.assertContained(section+":m2 read:test2", out)
+    self.assertContained(section+":m0 read m0:test0_0", out)
+
   def check_simd(self, expected_simds, expected_out):
     if SPIDERMONKEY_ENGINE in JS_ENGINES:
       out = run_js('a.out.js', engine=SPIDERMONKEY_ENGINE, stderr=PIPE, full_output=True)
